Add ingress.modSecurity.secRuleEngine support
In support of https://gitlab.com/gitlab-org/gitlab/8558, this
MR adds support for passing `modSecurity.secRuleEngine` to
customize execution of modsecurity within the ingress nginx
configuration.

This configuration is enabled as "DetectionOnly" by default within the
nginx-controller ConfigMap but overridable per ingress resource. This
allows customization of the specific handling of the ruleset, in this
case enabling the engine to be toggled to "Off" or "On" to either
disable rulechecks entirely or enable blocking mode.

This MR is paired with
https://gitlab.com/gitlab-org/cluster-integration/auto-deploy-image/merge_requests/28
which propagates the necessary configuration to the chart as part of
Auto Devops